+++ native/android/app/src/main/java/work/calmato/prestopay/MainActivity.kt
@@ -2,6 +2,7 @@ package work.calmato.prestopay
 
 import android.os.Bundle
 import android.util.Log
+import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.gms.ads.*
 import com.google.android.gms.ads.formats.NativeAdOptions
@@ -16,6 +17,32 @@ class MainActivity : AppCompatActivity() {
 
     // Initialize Mobile Ads SDK
     MobileAds.initialize(this)
+    refreshAd()
+  }
+  companion object {
+    var nativeAd : UnifiedNativeAd? = null
   }
 
+  fun refreshAd() {
+    // Native ad
+    val adLoader = AdLoader.Builder(this, "ca-app-pub-3940256099942544/2247696110")
+      .forUnifiedNativeAd { ad: UnifiedNativeAd ->
+        nativeAd = ad
+      }
+      .withAdListener(object : AdListener() {
+        // AdListener callbacks like OnAdFailedToLoad, OnAdOpened, OnAdClicked and
+        // so on, can be overridden here.
+        override fun onAdFailedToLoad(errorCode: Int) {
+          Log.i("MainActivity", "onAdFailedToLoad: $errorCode")
+        }
+      })
+      .withNativeAdOptions(
+        NativeAdOptions.Builder()
+          // Methods in the NativeAdOptions.Builder class can be
+          // used here to specify individual options settings.
+          .build()
+      )
+      .build()
+    adLoader.loadAd(AdRequest.Builder().build())
+  }
 }
