+++ native/android/app/src/main/java/work/calmato/prestopay/util/AdapterGroupPlane.kt
@@ -11,6 +11,7 @@ import com.google.android.gms.ads.AdLoader
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.formats.NativeAdOptions
 import com.google.android.gms.ads.formats.UnifiedNativeAd
+import work.calmato.prestopay.MainActivity
 import work.calmato.prestopay.databinding.ListGroupItemPlaneBinding
 import work.calmato.prestopay.network.GroupPropertyResponse
 
@@ -25,7 +26,7 @@ class AdapterGroupPlane(
       val rawData = value.toMutableList()
       if (!isHidden) {
         val size = value.size
-        val divideNum = 7
+        val divideNum = 6
         for (i in 1..size / divideNum) {
           rawData.add(
             i * divideNum, GroupPropertyResponse(
@@ -54,26 +55,9 @@ class AdapterGroupPlane(
   override fun onBindViewHolder(holder: AddGroupViewHolder, position: Int) {
     if (groupList[position].selected) {
       holder.binding.nativeAd.visibility = ImageView.VISIBLE
-      // Native ad
-      val adLoader = AdLoader.Builder(context, "ca-app-pub-3940256099942544/2247696110")
-        .forUnifiedNativeAd { ad: UnifiedNativeAd ->
-          holder.binding.nativeAd.setNativeAd(ad)
-        }
-        .withAdListener(object : AdListener() {
-          // AdListener callbacks like OnAdFailedToLoad, OnAdOpened, OnAdClicked and
-          // so on, can be overridden here.
-          override fun onAdFailedToLoad(errorCode: Int) {
-            Log.i("MainActivity", "onAdFailedToLoad: $errorCode")
-          }
-        })
-        .withNativeAdOptions(
-          NativeAdOptions.Builder()
-            // Methods in the NativeAdOptions.Builder class can be
-            // used here to specify individual options settings.
-            .build()
-        )
-        .build()
-      adLoader.loadAd(AdRequest.Builder().build())
+      if(MainActivity.nativeAd != null){
+        holder.binding.nativeAd.setNativeAd(MainActivity.nativeAd)
+      }
     } else {
       holder.binding.nativeAd.visibility = ImageView.GONE
       holder.binding.also {
